requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ChatKit & AI Agent Framework
openai-chatkit>=1.0.0
openai-agents>=0.2.0
# openai>=1.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Vector Search & Embeddings
qdrant-client>=1.7.0
cohere>=5.0.0

# Configuration & Environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Utilities
httpx>=0.26.0  # Better Auth API calls ke liye

# Development
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=4.1.0
# black>=24.0.0
# ruff>=0.1.0
# mypy>=1.8.0
